An experimental blog import from AGIRobots Blog
This shows how things are supposed to look like when we import posts
from AGI Robots Blog.

Source: https://developers.agirobots.com/jp/how-to-run-aloha/

Note: Beware that it's not something automated! I fetched a raw HTML
file from AGIRobots blog, convorted it to Markdown using pandoc, and
then fixed up here and there.

This is meant to be a demo, and we should use tools (such as wordpress
exporter) for production usage.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>